Remove BaseTechnology as overly complex.
