+++ undyingkingdoms/models/technologies/__init__.py
@@ -1,5 +1,1 @@
-# from .helpers import hoist_cols_macro
 from .technologies import Technology
-from .base_technology import BaseTechnology
-
-# hoist_cols_macro(Technology, BaseTechnology, sub_table_alias='base')
